+++ front/context/src/ApiRxContext.tsx
@@ -3,13 +3,19 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { ApiRx } from '@polkadot/api';
+import { DeriveFees } from '@polkadot/api-derive/types';
+import { EraIndex } from '@polkadot/types/interfaces';
 import { logger } from '@polkadot/util';
 import React, { useEffect, useState } from 'react';
+import { take } from 'rxjs/operators';
 
 import { ApiRxContextProviderProps } from './types';
+import { useDidUpdateEffect } from './util';
 
 export interface ApiRxContextType {
-  api: ApiRx; // From @polkadot/api\
+  api: ApiRx; // From @polkadot/api
+  bondingDuration?: EraIndex;
+  fees?: DeriveFees;
   isApiReady: boolean;
 }
 
@@ -24,9 +30,11 @@ export function ApiRxContextProvider(
 ): React.ReactElement {
   const { children = null, provider } = props;
   const [apiRx, setApiRx] = useState<ApiRx>(new ApiRx({ provider }));
+  const [bondingDuration, setBondingDuration] = useState<EraIndex>();
+  const [fees, setFees] = useState<DeriveFees>();
   const [isReady, setIsReady] = useState(false);
 
-  useEffect(() => {
+  useDidUpdateEffect(() => {
     // We want to fetch all the information again each time we reconnect. We
     // might be connecting to a different node, or the node might have changed
     // settings.
@@ -47,8 +55,26 @@ export function ApiRxContextProvider(
     return (): void => subscription.unsubscribe();
   }, [apiRx.isReady]);
 
+  useEffect(() => {
+    if (isReady) {
+      const subscription = apiRx.derive.balances
+        .fees()
+        .pipe(take(1))
+        .subscribe(derivedFees => {
+          setFees(derivedFees);
+        });
+
+      const duration = apiRx.consts.staking.bondingDuration;
+      setBondingDuration(duration);
+
+      return (): void => subscription.unsubscribe();
+    }
+  }, [isReady]);
+
   return (
-    <ApiRxContext.Provider value={{ api: apiRx, isApiReady: isReady }}>
+    <ApiRxContext.Provider
+      value={{ api: apiRx, bondingDuration, fees, isApiReady: isReady }}
+    >
       {children}
     </ApiRxContext.Provider>
   );