+++ front/gatsby/src/components/index.ts
@@ -2,11 +2,15 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+export * from './AccountsDropdown';
 export * from './AddressSummary';
 export * from './BondingModal';
 export * from './Button';
 export * from './Cart';
 export * from './Layout';
+export * from './NominationDetails';
 export * from './Seo';
+export * from './StatItem';
 export * from './Status';
 export * from './Table';
+export * from './Typography';