+++ front/gatsby/src/util/index.ts
@@ -4,3 +4,7 @@
 
 export * from './constants';
 export * from './cartHelpers';
+export * from './expectedRewardCalc';
+export * from './graphql';
+export * from './stakingHelpers';
+export * from './validateExtrinsic';